requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
groq>=0.4.0
scikit-learn>=1.4.0
numpy>=1.26.0
shap>=0.44.0
markdown>=3.5.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
